Update README.md added YT vid to show how to use
Yesterday I gave an impromptu session for our monthly gathering for FileMaker users at the Kentucky FileMaker developers meeting because the scheduled speaker that was to come cancel. Now because of that, I redid my promo video because it was all fresh in my head and I'm including it here to help you.